+++ Tests/CocoaPods/GRDBiOS-static/Extra.swift
@@ -4,10 +4,10 @@ func test_SQLITE_ENABLE_FTS5() {
     _ = FTS5.self
 }
 
-func test_SQLITE_ENABLE_SNAPSHOT() {
-    _ = DatabaseHistoricalSnapshot.self
-}
-
 func test_SQLITE_ENABLE_PREUPDATE_HOOK() {
     _ = DatabasePreUpdateEvent.self
+}
+
+func test_SQLITE_ENABLE_SNAPSHOT() {
+    _ = DatabaseHistoricalSnapshot.self
 }